requirements.txt
# Core AI/ML Libraries
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.24.0

# Music API Integration
spotipy>=2.22.0

# Vector Database & Similarity
chromadb>=0.4.0

# Web Interface & Visualization
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0

# Image Processing (for UI enhancements)
pillow>=10.0.0

# HTTP Requests
requests>=2.31.0

# Machine Learning
scikit-learn>=1.3.0

# Optional: Enhanced UI Components
streamlit-option-menu>=0.3.6
streamlit-lottie>=0.0.5
